requirements.txt
certifi==2017.4.17
chardet==3.0.4
click==6.7
enum-compat==0.0.2
enum34==1.1.6
eventlet==0.21.0
Flask==0.12.2
greenlet==0.4.12
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
python-engineio==1.7.0
python-socketio==1.7.6
requests==2.18.1
six==1.10.0
urllib3==1.21.1
Werkzeug==0.12.2
